requirements.txt
# These are required to run the tests.
Django==1.11
django-import-export==0.6.1

# Everything else is in a Django-version-free version
# for TravisCI.
-r travis.txt
